[acme] extract Disk
Move Disk into a separate Go file to mirror the source arrangement of
the original C code.
